adsp: passivate all occurrences of the qcadsp8280.mbn file
this is to ensure its not loaded from rootfs

Loading the adsp firmware leads to a short interruption on the
USB type-c VBUS line. A not self-powered device on that bus needs
to be re-detected then. If you've booted from type-c and you
have the rootfs already mounted, it is game over since the rootfs
is now inoperable.

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>