Adds clarification around explicitly invoking `nvidia-drm.modeset=1` as a kernel parameter.
In short, the Pop 24.04 Nvidia alpha ISOs required a fix for their
failure to reach GUI sessions on boot. While this is probably not
the "correct" solution, it is a functional one and meets needs in
time for relevent deadlines. This should be revisited more fully
after the first alpha releases, and a correct solution implemented.