Add loopback.cfg
Fixes: https://github.com/pop-os/iso/issues/243
( And the duplicate, https://github.com/pop-os/pop/issues/2708 )

I made the design choice to make the main grub.cfg in the iso have
support for being loaded "normally" or as a loopback.cfg.

It's the way that grml and NixOS support this, and it's how I
recommended that distros support it back in 2010 (😮) when I wrote
https://www.supergrubdisk.org/wiki/Loopback.cfg .

I think the reasoning on that page is still sound, and this also gets you
the benefit of being able to add your own extra kernel parameters before
loading the loopback.cfg.

For example, I like to boot live ISOs to RAM, and be able to safely
remove the device I booted from, while remaining booted.

I can do that by setting and exporting $extra_kernel_parameters before
loading the loopback.cfg, like so:

	extra_kernel_parameters="toram"
	export extra_kernel_parameters

	configfile /boot/grub/loopback.cfg

Ubuntu creates a completely separate grub.cfg and loopback.cfg, with
just 'iso-scan/filename=$iso_path' added as a kernel parameter in the
loopback.cfg, and everything else being identical.
( I assume both files are built from the same template file during
build, avoiding manual duplication, but I haven't checked to confirm )